reader/csv: sniff column name and type
The format is `COLUMN_NAME:COLUMN_TYPE`.

As part of this change, the CSV reader module has been refactored to use
"Drivers" for the actual storing of the values. This allows easy
implementation of sniffing, header skipping, and parsing, without
affecting performance (because it's all compile time) and it also keeps
the reader code clean.